feat: implement values for ir
To work with Calcite and Tarantool the planner needs to store values
in the plan leaf nodes. These values should be defined with exact types.
At the moment we need only several primitive types:
- boolean
- null
- number (decimal from decnumber library)
- string

As the planner analyzes SQL, it has to work with tree-valued logic
(3VL). At the moment we'll need only equivalence operation, so it
was implemented for the values in a 3VL manner.