refactoring!(c api): rename and move objects in the engine
BREAKING CHANGE: rename execute_query to dispatch_query

The main reason is that we'll soon need not only a coordinator
runtime, but also a new storage runtime as well. Current names
can confuse us in future.

Also add an ability to evaluate an arbitrary eviction funtion for
the LRU cache (on the storage side we'll need it to flush prepared
statements from the Tarantool cache when our LRU cache evicts the
one).

LRU cache was moved from within the cache folder into the engine.

Router logic was moved to a separate file in the cartridge module.