ci(cargo.lock): remove cargo.lock from .gitignore
We want to get reproducible builds. Cargo.lock does exactly this.
On the start of the development we were following the rule that
"libraries does not include cargo.lock in git, binary does".
However we misinterpreted the role of Sbroad. It is built as a "cdylib"
so we should treat it as an executable and include cargo.lock