fix: prepared statements on the storage
Previously we didn't take into account the fact, that prepared
statements act much alike OS descriptors. Though their internal
representation is located in the instance-wide memory of Tarantool,
we need to call "prepare" method to make its id from the global
cache for every new session. If the statement was not found in the
cache, then "prepare" compiles it and stores back to the cache.