feat: migrate to dynamic linking for unit tests
Previously, we created a static archives of the msgpuck and decNumber
libraries and made a statically linked them into the test executable.
After tarrantool module migrated to dlsym, we can no longer use static
linking.

As a result we build shared libraries for msgpuck and decNumber to
dynamic link them into the unit test binary.