feat: implement query statistics
The query statistics samples the execution time of the opentelemetry's
spans among 1% of all the running queries. Collected data is stored
in the temporary spaces on each instance independently and can be
analyzed by user (with the local SQL) or exported somewhere.
The statistics spaces would be created by the sbroad itself if they
do not exist yet.