feat: add statistics initialization to the API
Though we can lazily initialize statistics spaces, there is still a
problem with read-only replicas. If the workload does not write to
the storage (only reads with SELECTs), then callbre is used and no
queries will be dispatched to the storage master. As a result, no
statistics spaces will be initialized on the storage failover group.

To deal with it, we add a new API method to initialize statistics
on the instance startup.