feat: implement transformation undo log
As we use an append-only arena for the plan, we can easily implement
an UNDO log for each transformation we make. It can be helpful when
we would like to undo any of the transformations.

The main goal of it now - to rewind transformed selection to the
original state, as it can produce a logically equivalent, but too
long SQL string that breaks tarantool parser on the storages.