fix: trasaction already started error
Remove redundant transaction start from the open telemetry's spans
in the code. There is no sense in it at the moment (seems like it
was added while prototyping tarantool space as a storage for the
per-fiber traces... but we gave this idea up and switched to the
custom Rust implementation). But if we start a transaction manually
and call sbroad execute() method, we end up with transaction already
started error. So, remove the transaction start wrappers and make
nested transactions work.