feat(POC): picodata engine
Implement the picodata engine for sbroad to integrate distributed
SQL into the picodata binary. In fact, picodata engine simple
exports public rust functions from api.rs. On the picodata side
they (and some lua functions) are imported and included into the
bunary on the build phase.

Current commit is POC and have some problems:
- DDL is still not implemented in the picodata (_pico_space is mocked);
- we include tarantool module symbols twice (first time in sbroad,
  second time in picodata binary);

But anyway, it works and all these problems should be solved in the
next commits.