fix: block vshard rebalancing during dql
When reading from multiple storages it is
possible that some requests will execute faster
than the others. And it could lead to wrong
results when vhard moved the buckets between
nodes.

For example: we execute `select * from t`
on storage 1 and get the results. Vshard
moves the data from storage 1 to stroage 2,
and then we execute `select * from t` and in
result we get the data (that moved) twice.

This commit fixes it by using vshard's
storage ref api. Now multi-node read requests
first create the ref on each node which
blocks vshard rebalancing. Then we call
our stored procedure and remove the ref.