fix: re-implement trim
Previous implementation treated TRIM as an SQL built-in function,
but in reality it is a special expression in SQL. The main difference
is that "TRIM" is not a name but a reserved keyword, so it can't
be wrapped into double quotes. Moreover, if we treat it as a function,
we need some special hacks to handle a complicated logic in its
arguments. Now TRIM is a separated expression with a custom parsing
logic.