chore: make serde-yaml development dependency
Previous version was using yaml-rust which is marked as unmaintained.
This gets reported by dependency scanner we use for certification.
Though worth noting that serde-yaml itself is now archived. There is a
promising replacement library but no need to hurry with that just yet.
But anyway, we don't need this library in the production build, only
for tests. So, moved it to development dependencies.

Co-authored-by: Denis Smirnov <sd@picodata.io>