fix: public lua functions should fail with error
Previously some public lua functions in core-router.lua failed
returning nil instead of an error. As a result in a case of error
rust code treated it as success, but failed to pop expected results
from lua stack. So, instead of a real error user got lua reading
errors. Fixed.

Co-authored-by: Denis Smirnov <sd@picodata.io>