fix: to_params ignored repeating nodes
- to_params uses subtree iterator to create
parameters for cached query. But our plan
is a DAG, not a tree. We have multiple nodes
referring to the same node, this leads to
situation when we traverse the same subquery
node multiple times. This in term leads to
wrong number of parameters produces by
to_params
- fix this by keeping track of already traversed
parameter nodes in to_params